Remove some reduntant cast warning in product component (OFBIZ-4102)
git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1081070 13f79535-47bb-0310-9956-ffa450edef68

+++ applications/order/src/org/ofbiz/order/shoppingcart/ShoppingCartItem.java
@@ -1429,7 +1429,7 @@ public class ShoppingCartItem implements java.io.Serializable {
             BigDecimal quantityUsed = entry.getValue();
             if (productPromoId.equals(productPromoCondActionPK.getString("productPromoId")) && productPromoRuleId.equals(productPromoCondActionPK.getString("productPromoRuleId"))) {
                 entryIter.remove();
-                BigDecimal existingValue = (BigDecimal) this.quantityUsedPerPromoActual.get(productPromoCondActionPK);
+                BigDecimal existingValue = this.quantityUsedPerPromoActual.get(productPromoCondActionPK);
                 if (existingValue == null) {
                     this.quantityUsedPerPromoActual.put(productPromoCondActionPK, quantityUsed);
                 } else {
